+++ web/apps/photos/src/services/userService.ts
@@ -1,8 +1,7 @@
 import { putAttributes } from "@/accounts/api/user";
 import log from "@/base/log";
-import { apiURL, familyAppOrigin } from "@/base/origins";
+import { apiURL } from "@/base/origins";
 import type { UserDetails } from "@/new/photos/services/user";
-import { getLocalFamilyData, isPartOfFamily } from "@/new/photos/services/family";
 import { ApiError } from "@ente/shared/error";
 import HTTPService from "@ente/shared/network/HTTPService";
 import { LS_KEYS, getData } from "@ente/shared/storage/localStorage";
@@ -22,24 +21,6 @@ export const getPublicKey = async (email: string) => {
         },
     );
     return resp.data.publicKey;
-};
-
-export const getFamiliesToken = async () => {
-    try {
-        const token = getToken();
-
-        const resp = await HTTPService.get(
-            await apiURL("/users/families-token"),
-            null,
-            {
-                "X-Auth-Token": token,
-            },
-        );
-        return resp.data["familiesToken"];
-    } catch (e) {
-        log.error("failed to get family token", e);
-        throw e;
-    }
 };
 
 export const isTokenValid = async (token: string) => {
@@ -96,19 +77,6 @@ export const getUserDetailsV2 = async (): Promise<UserDetails> => {
         return resp.data;
     } catch (e) {
         log.error("failed to get user details v2", e);
-        throw e;
-    }
-};
-
-export const getFamilyPortalRedirectURL = async () => {
-    try {
-        const jwtToken = await getFamiliesToken();
-        const isFamilyCreated = isPartOfFamily(getLocalFamilyData());
-        return `${familyAppOrigin()}?token=${jwtToken}&isFamilyCreated=${isFamilyCreated}&redirectURL=${
-            window.location.origin
-        }/gallery`;
-    } catch (e) {
-        log.error("unable to generate to family portal URL", e);
         throw e;
     }
 };
